+++ CHANGELOG.md
@@ -5,6 +5,13 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Modify the `sources` query to return results that also include
+  sources from its peers.
+
 ## [0.18.0]
 
 ### Changed
@@ -439,6 +446,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Initial release.
 
+[Unreleased]: <https://github.com/aicers/giganto/compare/0.18.0...main>
 [0.18.0]: <https://github.com/aicers/giganto/compare/0.17.0...0.18.0>
 [0.17.0]: <https://github.com/aicers/giganto/compare/0.16.0...0.17.0>
 [0.16.0]: <https://github.com/aicers/giganto/compare/0.15.3...0.16.0>